requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-pptx==0.6.22
python-multipart==0.0.12
